refactor(DEX-4034): sidebar was moved to a new component and name changes.
